Fix a number of issues with remote commands
It turns out the RHEL 7.1 ControlMaster problem I was having is
likely that ssh is dying.  We weren't checking for return codes
though, so the fact that ssh failed was being ignored and bogus
results passed to the widgets.  This stops that.

It also fixes up the logic around retrieving the file contents.
There's no reason to be making multiple calls that may or may not
start the retrieval process, and having it work that way makes it
difficult to reason about what is going on.  Be more explicit about
which calls start the retrieval and which just want to know the
local file path.

This doesn't solve the ssh bug, but it does at least allow us to
gracefully degrade in those cases.