Improvements to remote host support
Add Username and IDFile parameter for changing those when ssh'ing
to a remote system.

Also disable host key checking as it is incredibly annoying to
have a host key popup dozens of times because you reprovisioned a
system (it's also quite difficult to fix because at least on my
systems the host key input steals keyboard focus so you can't type
in a terminal to verify the host key).

Finally, fixes a bug with the memory widget where a SubType setting
of Graph was not honored.