Make window fixed size
This is necessary to allow the window to be placed correctly in the
strut when DE's like KDE are set up to maximize a window when it gets
moved to the edge of the screen. Making the window fixed size prevents
that behavior and leaves the window where it is dropped.

A few things are also cleaned up in the X property setting code.